+++ src/util/StringUtil.cpp
@@ -324,7 +324,7 @@ void StringUtil::parse_table( const char * str, int len, std::vector < std::vect
 
     table.resize( lines.size() );
     int ncol = 0;
-    for ( int i = 0; i < lines.size(); i++ )
+    for ( int i = 0; i < (int)lines.size(); i++ )
     {
         // strtok won't work on const char *, so make a copy.
         char buf[ lines[i].size() ];
@@ -338,14 +338,14 @@ void StringUtil::parse_table( const char * str, int len, std::vector < std::vect
             f = strtok ( nullptr, "\t" );
         }
 
-        if ( table[i].size() > ncol )
+        if ( (int)table[i].size() > ncol )
         {
             ncol = table[i].size();
         }
     }
 
     // ensure all rows have same number of columns.
-    for ( int i = 0; i < table.size(); i++ )
+    for ( int i = 0; i < (int)table.size(); i++ )
     {
         table[i].resize( ncol );
     }
